# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Foxy P.A. APIs",
        "description": "Foxy P.A. helps users organize and manage appointments by allowing them to search for clinics, follow doctors, set appointments, and receive notifications. It provides various time management modes, lets users add titles and descriptions to appointment requests, and enables anyone signed in to set or receive appointments.",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/requests/setrequest": {
            "post": {
                "tags": [
                    "Request"
                ],
                "summary": "Set up a request",
                "description": "Sets up a request and returns necessary information about the user and their availability.",
                "operationId": "setRequest",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "other_user"
                                ],
                                "properties": {
                                    "other_user": {
                                        "description": "ID of the other user",
                                        "type": "integer",
                                        "example": 2
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request information retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "userId": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "info": {
                                            "description": "User's work hours and time zone",
                                            "properties": {
                                                "work_from": {
                                                    "type": "string",
                                                    "example": "08:00:00"
                                                },
                                                "work_to": {
                                                    "type": "string",
                                                    "example": "17:00:00"
                                                },
                                                "time_zone": {
                                                    "type": "string",
                                                    "example": "UTC"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "full_days": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "2023-10-21"
                                            }
                                        },
                                        "info_day": {
                                            "description": "User's work days",
                                            "properties": {
                                                "monday": {
                                                    "type": "string",
                                                    "example": "accepted"
                                                },
                                                "tuesday": {
                                                    "type": "string",
                                                    "example": "accepted"
                                                },
                                                "wednesday": {
                                                    "type": "string",
                                                    "example": "accepted"
                                                },
                                                "thursday": {
                                                    "type": "string",
                                                    "example": "accepted"
                                                },
                                                "friday": {
                                                    "type": "string",
                                                    "example": "accepted"
                                                },
                                                "saturday": {
                                                    "type": "string",
                                                    "example": "closed"
                                                },
                                                "sunday": {
                                                    "type": "string",
                                                    "example": "closed"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "An error occurred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter your bearer token in the format **Bearer &lt;token&gt;**",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}